fix(clouds): 구름이 계단으로 움직이는 문제 — half 정밀도 상수를 float 로
구름이 부드럽게 흐르지 않고 픽셀 단위로 덜컥거리는 문제.

원인은 패키지가 구름 위치 누적값을 half 로 선언한 것이다. half 는
값이 커질수록 표현 간격이 벌어져 1024 근처에서 간격이 1.0 이나 된다.
이 씬은 드리프트가 아주 느려서(DriftSpeed 0.025) 한 프레임 이동량이
그 간격보다 훨씬 작다. 오프셋이 한동안 그대로 있다가 어느 순간 한 칸
튀니 구름이 계단으로 움직인다.

속도를 낮춘 뒤에 도드라진 이유도 이것이다. 빠르면 매 프레임 표현
간격을 넘어가서 오히려 티가 덜 난다.

VolumetricCloudsDefs.hlsl 에서 half -> float:
  _ShapeNoiseOffset              구름 형상 흐름 오프셋
  _VerticalShapeNoiseOffset      수직 형상 오프셋
  _WindVector                    바람에 의한 위치 변위
  _VerticalShapeWindDisplacement 수직 형상 바람 변위
  _VerticalErosionWindDisplacement 수직 침식 바람 변위
  _AccumulationFactor            재투영 블렌드 가중치

효과 (ProbeShimmer, 구름만 흐르는 조건의 프레임 간 변화):
  시간누적 0.5    1.99 -> 0.76
  전체 해상도     1.91 -> 0.70
같은 이동량에 화면 변화가 2.6 배 줄었다 = 계단이 부드러워졌다.

디노이즈 패스의 화면 좌표도 half2 -> float2 로 올렸다(prevPosCS,
curPosCS, velocity). 다만 이쪽은 측정상 효과가 없었다 — 카메라 이동
시 깜빡임 6.2% 로 그대로였다. 정밀도상 맞는 방향이라 유지하되,
카메라 이동 쪽 떨림의 원인은 아니다.

CBUFFER 안에서 half 는 어차피 32 비트 슬롯을 쓰므로 상수 버퍼
레이아웃과 C# 대입 코드는 영향 없다.

앞서 고친 아지랑이(디더링)와는 다른 문제다. 시간누적은 두 증상 모두와
무관하다 — 0/0.5/1 전부 같았다.

THIRD_PARTY.md 에 패치 2 항목으로 기록.

+++ unity/DRON/Assets/VolumetricClouds/VolumetricCloudsDefs.hlsl
@@ -8,12 +8,29 @@ half _NumLightSteps;
 half _MaxStepSize;
 float _HighestCloudAltitude;
 float _LowestCloudAltitude;
-half4 _ShapeNoiseOffset;
-half _VerticalShapeNoiseOffset;
+// [여의도63 수정] half -> float. 구름 흐름 오프셋이다.
+//
+//   half 는 값이 커질수록 표현 간격이 벌어진다. 1024 근처에서는 간격이
+//   1.0 이나 된다. 이 씬의 드리프트는 아주 느려서(0.025) 한 프레임
+//   이동량이 그 간격보다 훨씬 작다. 그러면 오프셋이 한동안 그대로
+//   있다가 어느 순간 한 칸 튀고, 구름이 미끄러지는 대신 계단으로
+//   움직인다 — "움직이면서 드르륵" 의 직접 원인이다.
+//
+//   느린 속도일수록 심해진다. 빠르면 매 프레임 표현 간격을 넘어가서
+//   오히려 덜 티가 난다. 속도를 낮춘 뒤에 도드라진 이유가 이것이다.
+//
+//   CBUFFER 안에서 half 는 어차피 32 비트 슬롯을 차지하므로
+//   레이아웃과 C# 쪽 대입은 영향 없다.
+float4 _ShapeNoiseOffset;
+float _VerticalShapeNoiseOffset;
 half4 _WindDirection;
-half4 _WindVector;
-half _VerticalShapeWindDisplacement;
-half _VerticalErosionWindDisplacement;
+// [여의도63 수정] half -> float. 바람이 구름 위치를 미는 누적 변위다.
+// _ShapeNoiseOffset 과 같은 이유로 half 면 이동이 계단이 된다.
+// (VolumetricCloudsUtilities.hlsl:338, :345 에서 위치에 더해진다)
+float4 _WindVector;
+// [여의도63 수정] half -> float. 위와 같은 누적 변위다.
+float _VerticalShapeWindDisplacement;
+float _VerticalErosionWindDisplacement;
 half _MediumWindSpeed;
 half _SmallWindSpeed;
 half _AltitudeDistortion;
@@ -33,7 +50,10 @@ half4 _ScatteringTint;
 half _AmbientProbeDimmer;
 half _SunLightDimmer;
 float _EarthRadius;
-half _AccumulationFactor;
+// [여의도63 수정] half -> float. 재투영 블렌드 가중치다.
+// 디노이즈 패스에서 velocity(float)와 같이 쓰이는데 여기서 half 로
+// 떨어지면 그 정밀도가 도로 깎인다. 레이아웃은 안 변한다.
+float _AccumulationFactor;
 half _NormalizationFactor;
 half _CloudNearPlane;
 CBUFFER_END
